moved open ai specific utils to a separate file
