+++ crates/goose/src/providers.rs
@@ -7,6 +7,7 @@ pub mod model_pricing;
 pub mod oauth;
 pub mod ollama;
 pub mod openai;
+pub mod openai_utils;
 pub mod utils;
 
 pub mod google;
